+++ modules/router/src/config.ts
@@ -1,4 +1,4 @@
-import { TAddress, TChainId, TIntegerString, TDecimalString } from "@connext/vector-types";
+import { TUrl, TChainId, TAddress, TIntegerString, AllowedSwapSchema } from "@connext/vector-types";
 import { Static, Type } from "@sinclair/typebox";
 import Ajv from "ajv";
 import { getAddress } from "@ethersproject/address";
@@ -15,22 +15,12 @@ const RebalanceProfileSchema = Type.Object({
 });
 export type RebalanceProfile = Static<typeof RebalanceProfileSchema>;
 
-const AllowedSwapSchema = Type.Object({
-  fromChainId: TChainId,
-  toChainId: TChainId,
-  fromAssetId: TAddress,
-  toAssetId: TAddress,
-  priceType: Type.Union([Type.Literal("hardcoded")]),
-  hardcodedRate: TDecimalString,
-});
-export type AllowedSwap = Static<typeof AllowedSwapSchema>;
-
 const VectorRouterConfigSchema = Type.Object({
   adminToken: Type.String(),
   allowedSwaps: Type.Array(AllowedSwapSchema),
-  chainProviders: Type.Dict(Type.String({ format: "uri" })),
-  dbUrl: Type.Optional(Type.String({ format: "uri" })),
-  nodeUrl: Type.String({ format: "uri" }),
+  chainProviders: Type.Dict(TUrl),
+  dbUrl: Type.Optional(TUrl),
+  nodeUrl: TUrl,
   logLevel: Type.Optional(
     Type.Union([
       Type.Literal("fatal"),
@@ -42,20 +32,26 @@ const VectorRouterConfigSchema = Type.Object({
       Type.Literal("silent"),
     ]),
   ),
+  messagingUrl: Type.Optional(TUrl),
   rebalanceProfiles: Type.Array(RebalanceProfileSchema),
+  mnemonic: Type.Optional(Type.String()),
 });
 
 type VectorRouterConfig = Static<typeof VectorRouterConfigSchema>;
 const dbUrl = process.env.VECTOR_DATABASE_URL;
 let vectorConfig: VectorRouterConfig;
+
+const mnemonicEnv = process.env.VECTOR_MNEMONIC;
 try {
   vectorConfig = JSON.parse(process.env.VECTOR_CONFIG!);
 } catch (e) {
   throw new Error(`VECTOR_CONFIG contains invalid JSON: ${e.message}`);
 }
+const mnemonic = mnemonicEnv || vectorConfig.mnemonic;
 
 // Set defaults
 vectorConfig.nodeUrl = vectorConfig.nodeUrl || "http://node:8000";
+vectorConfig.messagingUrl = vectorConfig.messagingUrl || "http://messaging";
 
 const validate = ajv.compile(VectorRouterConfigSchema);
 const valid = validate(vectorConfig);
@@ -90,4 +86,5 @@ vectorConfig.rebalanceProfiles = vectorConfig.rebalanceProfiles.map((profile) =>
 export const config = {
   dbUrl,
   ...vectorConfig,
-} as VectorRouterConfig;
+  mnemonic,
+} as Omit<VectorRouterConfig, "mnemonic"> & { mnemonic: string };