+++ modules/router/src/router.ts
@@ -4,6 +4,7 @@ import { Registry } from "prom-client";
 
 import { setupListeners } from "./listener";
 import { IRouterStore } from "./services/store";
+import { IRouterMessagingService } from "./services/messaging";
 
 export interface IRouter {
   startup(): Promise<void>;
@@ -16,6 +17,7 @@ export class Router implements IRouter {
     private readonly nodeService: INodeService,
     private readonly chainReader: IVectorChainReader,
     private readonly store: IRouterStore,
+    private readonly messagingService: IRouterMessagingService,
     private readonly logger: BaseLogger,
     private readonly register: Registry,
   ) {}
@@ -26,22 +28,34 @@ export class Router implements IRouter {
     nodeService: INodeService,
     chainReader: IVectorChainReader,
     store: IRouterStore,
+    messagingService: IRouterMessagingService,
     logger: BaseLogger,
     register: Registry,
   ): Promise<Router> {
-    const router = new Router(publicIdentifier, signerAddress, nodeService, chainReader, store, logger, register);
+    const router = new Router(
+      publicIdentifier,
+      signerAddress,
+      nodeService,
+      chainReader,
+      store,
+      messagingService,
+      logger,
+      register,
+    );
     await router.startup();
     logger.info("Vector Router connected 🚀");
     return router;
   }
 
   async startup(): Promise<void> {
+    await this.messagingService.connect();
     await setupListeners(
       this.publicIdentifier,
       this.signerAddress,
       this.nodeService,
       this.store,
       this.chainReader,
+      this.messagingService,
       this.logger,
       this.register,
     );