+++ modules/documentation/docs/changelog.md
@@ -2,6 +2,38 @@
 
 ## Next Release
 
+## 0.1.6
+
+- \[browser-node\] Save crossChainTransfer error state
+
+## 0.1.5
+
+- \[browser-node\] Fix bug in `crossChainTransferStore` key
+
+## 0.1.4
+
+- \[contracts\] Fix bug in `chainService`
+
+## 0.1.3
+
+- \[browser-node\] Remove from storage if transfer is cancelled.
+- \[browser-node\] Register listener for sender cancellations once
+- \[browser-node\] Don't hang on `crossChainTransfer` if sender-side transfer cancelled
+- \[router\] Handle unverified (sent a single-signed) updates for transfer forwarding on check-in (using `reconcileDeposit` for state consensus)
+- \[contracts\] Improvements with `gasPrice` and `gasLimit` in chain servce
+
+## 0.1.3-beta.1
+
+- \[browser-node\] Remove from storage if transfer is cancelled.
+
+## 0.1.2
+
+- \[router\] Add NATS client
+- \[browser-node, server-node\] Add methods to get router config
+- \[browser-node\] Verify swap supported before `crossChainTransfer`
+- \[router\] Hotfix `toString()` undefined
+- \[ops\] Add LogDNA tag support to router config
+
 ## 0.1.1
 
 - \[all\] Improve logging